requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# HTTP Requests
requests==2.31.0

# Data Processing
numpy==1.24.3
Pillow==10.0.1

# Development
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
